+++ src/entities/Product/index.ts
@@ -3,6 +3,5 @@ export { ProductDetails } from "./ui/ProductDetails/ProductDetails"
 export { ProductCardBasket } from "./ui/ProductCardBasket/ProductCardBasket"
 
 export { ProductCardProps } from "./ui/ProductCard/ProductCard"
-export { ProductDetailsSchema } from "./model/types/productDetailsSchema"
-export { productDetailsReducer } from "./model/slice/productDetailsSlice"
-export type { Product } from "./model/types/productDetailsSchema"
+
+export type { Product } from "./model/types"
