+++ src/entities/Product/model/types.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    id?: number
+    name?: string
+    is_new?: boolean
+    images?: string[]
+    price?: number
+    currency?: string
+}
