--- src/entities/Product/model/types/productDetailsSchema.ts
@@ -1,13 +1,0 @@
-export interface Product {
-    id: number
-    name: string
-    is_new: boolean
-    images: string[]
-    price: number
-}
-
-export interface ProductDetailsSchema {
-    data?: Product
-    isLoading: boolean
-    error?: string
-}
