+++ src/pages/ProductDetailsPage/index.ts
@@ -1,3 +1,4 @@
-import { ProductDetailsPage } from "./ui/ProductDetailsPage"
+export { ProductDetailsPage } from "./ui/ProductDetailsPage"
 
-export { ProductDetailsPage }
+export { ProductDetailsSchema } from "./model/types/productDetailsSchema"
+export { productDetailsReducer } from "./model/slice/productDetailsSlice"
