+++ src/pages/ProductDetailsPage/model/selectors/productDetailsSelectors.ts
