+++ src/pages/ProductDetailsPage/model/slice/productDetailsSlice.ts
