+++ src/pages/ProductDetailsPage/model/types/productDetailsSchema.ts
@@ -1,0 +1,7 @@
+import { Product } from "entities/Product"
+
+export interface ProductDetailsSchema {
+    data?: Product
+    isLoading: boolean
+    error?: string
+}
