+++ src/pages/ProductDetailsPage/ui/ProductDetailsPage.tsx
@@ -1,9 +1,20 @@
+import { useEffect } from "react"
+import { useParams } from "react-router-dom"
 import { Breadcrumbs } from "entities/Breadcrumbs"
 import { ProductDetails } from "entities/Product"
-import { useDispatch } from "react-redux"
+import { useDispatch, useSelector } from "react-redux"
 import { AppRoutes } from "shared/config/routeConfig/routeConfig"
 import { BannersRow } from "widgets/BannersRow"
 import { ProductCarousel, ProductCarouselVariant } from "widgets/ProductCarousel"
+import { fetchProductDetails } from "../model/services/fetchProductDetails/fetchProductDetails"
+import {
+    getProductDetailsError,
+    getProductDetailsId,
+    getProductDetailsImages,
+    getProductDetailsLoading,
+    getProductDetailsName,
+    getProductDetailsPrice,
+} from "../model/selectors/productDetailsSelectors"
 
 export function ProductDetailsPage() {
     const breadcrumbsList = [
@@ -13,10 +24,35 @@ export function ProductDetailsPage() {
         AppRoutes.PRODUCT_DETAILS,
     ]
 
+    const dispatch = useDispatch()
+    const { id } = useParams()
+
+    useEffect(() => {
+        dispatch(fetchProductDetails(id))
+    }, [dispatch, id])
+
+    const productRequestLoading = useSelector(getProductDetailsLoading)
+    const productRequestError = useSelector(getProductDetailsError)
+    const productId = useSelector(getProductDetailsId)
+    const productName = useSelector(getProductDetailsName)
+    const productPrice = useSelector(getProductDetailsPrice)
+    const productCurrency = "$" // to be changed
+    const prodductIsNew = false // to be changed
+    const productImages = useSelector(getProductDetailsImages) // to be used
+
     return (
         <div>
             <Breadcrumbs breadcrumbsList={breadcrumbsList} />
-            <ProductDetails />
+            <ProductDetails
+                isLoading={productRequestLoading}
+                error={productRequestError}
+                is_new={prodductIsNew}
+                id={productId}
+                name={productName}
+                price={productPrice}
+                currency={productCurrency}
+                images={productImages}
+            />
             <ProductCarousel variant={ProductCarouselVariant.TOP_PRODUCTS} />
             <BannersRow />
         </div>
